requirements.txt
asgiref==3.3.1
braintree==4.5.0
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.5
django-import-export==2.5.0
django-paypal==1.0.0
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.6
paypal==1.2.5
Pillow==8.1.0
python-decouple==3.4
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.2
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
